+++ packages/postgres-database/src/simcore_postgres_database/migration/versions/e84904303e2b_adding_groups_extra_properties_table.py
@@ -1,4 +1,4 @@
-"""adding internet to groups permissions
+"""adding groups extra properties table
 
 Revision ID: e84904303e2b
 Revises: 10b293fdcd56
@@ -18,7 +18,7 @@ depends_on = None
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
     op.create_table(
-        "internet_to_groups",
+        "groups_extra_properties",
         sa.Column("group_id", sa.BigInteger(), nullable=False),
         sa.Column(
             "has_access", sa.Boolean(), server_default=sa.text("true"), nullable=False
@@ -43,5 +43,5 @@ def upgrade():
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("internet_to_groups")
+    op.drop_table("groups_extra_properties")
     # ### end Alembic commands ###
