+++ services/director-v2/src/simcore_service_director_v2/modules/db/tables.py
@@ -2,7 +2,9 @@ from simcore_postgres_database.models.comp_pipeline import comp_pipeline
 from simcore_postgres_database.models.comp_runs import comp_runs
 from simcore_postgres_database.models.comp_tasks import NodeClass, StateType, comp_tasks
 from simcore_postgres_database.models.groups import user_to_groups
-from simcore_postgres_database.models.internet_to_groups import internet_to_groups
+from simcore_postgres_database.models.groups_extra_properties import (
+    groups_extra_properties,
+)
 from simcore_postgres_database.models.projects import ProjectType, projects
 from simcore_postgres_database.models.projects_networks import projects_networks
 
@@ -10,7 +12,7 @@ __all__ = [
     "comp_pipeline",
     "comp_runs",
     "comp_tasks",
-    "internet_to_groups",
+    "groups_extra_properties",
     "NodeClass",
     "projects_networks",
     "projects",
